# src/schemas/product.entity.ts
export interface ProductEntity {
  id: string;
  title: string;
  description: string;
  price: number;
}

export interface ProductModel {
  id: number;
  title: string;
  description: string;
  price: number;
}